+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v0.0.2
+
+- Add option for obtaining `QPASS_WORDS` from optional user's environment
+  variable settings.
+- Add runtime and help information to include usage of `--allow-env=QPASS_WORDS`
+  execution permission.
+- Show 'default' or 'user defined' for `QPASS_WORDS` number in `-a` or `--about`
+  output.
+
 ## v0.0.1
 
 - initial commit with basic file creations and layout